course_chat_bot/requirements.txt
python-telegram-bot
requests
pymongo